requirements.txt
dashscope
sounddevice
numpy
pynput
pyobjc-framework-Quartz
